requirements.txt
Flask
Flask-Login
Flask-Mail
flask_wtf
sqlalchemy
wtforms
pillow
qrcode